Decouple CommInitializer from CommChannelService
