+++ Opel-Manager/AndroidManifest.xml
@@ -29,7 +29,7 @@
         android:icon="@drawable/opel"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
-        <service android:name="com.opel.cmfw.view.CommService"/>
+        <service android:name="com.opel.cmfw.view.CommChannelService"/>
 
         <activity
             android:name="com.opel.opel_manager.view.MainActivity"
