+++ Opel-Manager/src/com/opel/cmfw/view/CommBroadcastReceiver.java
@@ -77,48 +77,3 @@ public class CommBroadcastReceiver extends BroadcastReceiver {
         }
     }
 }
-
-class CommBroadcaster {
-    public static void onInitializationResult(CommService service, boolean
-            isInitSuccess) {
-        Intent broadcastIntent = new Intent();
-        broadcastIntent.setAction(CommBroadcastReceiver.ACTION);
-        broadcastIntent.putExtra(CommBroadcastReceiver.KEY_EVENT_TYPE,
-                CommBroadcastReceiver.EVENT_TYPE_ON_INITIALIZATION_RESULT);
-        broadcastIntent.putExtra(CommBroadcastReceiver.KEY_IS_INIT_SUCCESS,
-                isInitSuccess);
-        service.sendBroadcast(broadcastIntent);
-    }
-
-    public static void onWifiStateChanged(CommService service, boolean
-            isWifiOn) {
-        Intent broadcastIntent = new Intent();
-        broadcastIntent.setAction(CommBroadcastReceiver.ACTION);
-        broadcastIntent.putExtra(CommBroadcastReceiver.KEY_EVENT_TYPE,
-                CommBroadcastReceiver.EVENT_TYPE_ON_WIFI_STATE_CHANGED);
-        broadcastIntent.putExtra(CommBroadcastReceiver.KEY_IS_WIFI_ON,
-                isWifiOn);
-        service.sendBroadcast(broadcastIntent);
-    }
-
-    public static void onBluetoothStateChanged(CommService service, boolean
-            isConnected) {
-        Intent broadcastIntent = new Intent();
-        broadcastIntent.setAction(CommBroadcastReceiver.ACTION);
-        broadcastIntent.putExtra(CommBroadcastReceiver.KEY_EVENT_TYPE,
-                CommBroadcastReceiver.EVENT_TYPE_ON_BLUETOOTH_STATE_CHANGED);
-        broadcastIntent.putExtra(CommBroadcastReceiver.KEY_IS_CONNECTED,
-                isConnected);
-        service.sendBroadcast(broadcastIntent);
-    }
-
-    public static void onReceivedMessage(CommService service, String message) {
-        Intent broadcastIntent = new Intent();
-        broadcastIntent.setAction(CommBroadcastReceiver.ACTION);
-        broadcastIntent.putExtra(CommBroadcastReceiver.KEY_EVENT_TYPE,
-                CommBroadcastReceiver.EVENT_TYPE_ON_RECEIVED_MESSAGE);
-        broadcastIntent.putExtra(CommBroadcastReceiver.KEY_MESSAGE,
-                message);
-        service.sendBroadcast(broadcastIntent);
-    }
-}
